# api-docs.yaml
openapi: 3.0.1
info:
  title: iBoard 社區公佈欄(後端Spring boot)
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Generated server url
...
paths:
...
  /rfids/{rfidId}:
    get:
      tags:
      - RFID CRUD
      summary: 取得RFID資料
      description: 透過rfid_id取的資料並回傳前端
      operationId: getRfidById
      parameters:
      - name: rfidId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Rfid'
    put:
      tags:
      - RFID CRUD
      summary: 修改RFID資料
      description: 透過rfid_id修改RFID資料數據
      operationId: updateRfid
      parameters:
      - name: rfidId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        description: 請求參數列表
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RfidRequest'
        required: true
      responses:
        "404":
          description: 無此 RFID ID，無法執行修改功能
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Rfid'
    delete:
      tags:
      - RFID CRUD
      summary: 刪除RFID資料
      description: 透過rfid_id刪除RFID資料數據
      operationId: deleteRfid
      parameters:
      - name: rfidId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: 成功
  /packages/{packageId}:
    get:
      tags:
      - 住戶包裹 CRUD
      summary: 取得包裹資料
      description: 透過package_id取得包裹資料並回傳前端
      operationId: getPackageById
      parameters:
      - name: packageId
        in: path
        description: 包裹id
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Package'
        "404":
          description: 無此 包裹 ID，無法執行查詢功能
    put:
      tags:
      - 住戶包裹 CRUD
      summary: 修改包裹資料(修改包裹為已領取狀態)
      description: 透過 package_id 修改包裹資料，透過RequestParam String state調整包裹狀態(即領取包裹)
      operationId: updatePackage
      parameters:
      - name: packageId
        in: path
        description: 包裹id
        required: true
        schema:
          type: integer
          format: int32
      - name: state
        in: query
        description: 包裹狀態(已領取/未領取)
        required: false
        schema:
          type: string
          default: "0"
      requestBody:
        description: 請求參數列表
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PackageRequest'
        required: true
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Package'
        "404":
          description: 無此 包裹 ID，無法執行修改功能
    delete:
      tags:
      - 住戶包裹 CRUD
      summary: 刪除包裹
      description: 刪除包裹資料
      operationId: deletePackage
      parameters:
      - name: packageId
        in: path
        description: 包裹id
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: 成功
...
  /rfids:
    get:
      tags:
      - RFID CRUD
      summary: 取的RFID資料列表(模糊查詢地址、RFID CODE)
      description: 可模糊查詢地址與RFID CODE，取得RFID資料列表
      operationId: getRfidList
      parameters:
      - name: address
        in: query
        description: RFID擁有者地址
        required: false
        schema:
          type: string
      - name: rfid
        in: query
        description: RFID
        required: false
        schema:
          type: string
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Rfid'
    post:
      tags:
      - RFID CRUD
      summary: 新增RFID資料
      description: 新增RFID資料並回傳新增的RFID資料數據
      operationId: createRfid
      requestBody:
        description: 請求參數列表
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RfidRequest'
        required: true
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Rfid'
  /packages:
    get:
      tags:
      - 住戶包裹 CRUD
      summary: 查詢包裹列表(模糊查詢地址+區隔領取與未領取+分頁+排序)
      description: 透過模糊查詢地址取得包裹資料列表；透過RFID CODE來抓取地址(一個住址可能有多個RFID，所以我只要使用其中一個RFID，就可以查到自己與同住家人的包裹；)，可選擇查詢領取或未領取(預設未領取)；依照前端請求orderBy指定項目排序(預設為預定日期時間)；依照前端請求sort指定排序順序(預設DESC由新到舊排序)；依照前端請求limit限制每頁筆數(預設5)；依照前端請求offset設定跳過筆數(預設0，代表第一頁)；回傳總數count讓前端計算分頁。
      operationId: getPackages
      parameters:
      - name: address
        in: query
        description: 住戶地址
        required: false
        schema:
          type: string
      - name: rfid
        in: query
        description: 住戶RFID CODE(當使用此方法時，會在業務邏輯強制把address設為null)
        required: false
        schema:
          type: string
      - name: state
        in: query
        description: 包裹狀態(未領取/已領取)
        required: false
        schema:
          type: string
          default: "0"
      - name: orderBy
        in: query
        description: 排序對象
        required: false
        schema:
          type: string
          default: last_modified_date
      - name: sort
        in: query
        description: 排序順序
        required: false
        schema:
          type: string
          default: DESC
      - name: limit
        in: query
        description: 限制資料筆數
        required: false
        schema:
          maximum: 20
          minimum: 0
          type: integer
          format: int32
          default: 5
      - name: offset
        in: query
        description: 跳過資料筆數
        required: false
        schema:
          minimum: 0
          type: integer
          format: int32
          default: 0
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Package'
    post:
      tags:
      - 住戶包裹 CRUD
      summary: 新增包裹
      description: 新增包裹資料
      operationId: createPackage
      parameters:
      - name: state
        in: query
        description: 包裹狀態(已領取/未領取)
        required: false
        schema:
          type: string
          default: "0"
      requestBody:
        description: 請求參數列表
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PackageRequest'
        required: true
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Package'
...
components:
  schemas:
...
    RfidRequest:
      required:
      - rfidCode
      - rfidName
      - userAddress
      type: object
      properties:
        userAddress:
          type: string
          description: 住戶地址
        rfidName:
          type: string
          description: RFID使用者
        rfidCode:
          type: string
          description: RFID編號
    Rfid:
      type: object
      properties:
        rfidId:
          type: integer
          description: RFID流水號
          format: int32
        userAddress:
          type: string
          description: 住戶地址
        rfidName:
          type: string
          description: RFID擁有者
        rfidCode:
          type: string
          description: RFID ID
        createdDate:
          type: string
          description: RFID 資料創建時間
          format: date-time
        lastModifiedDate:
          type: string
          description: RFID 資料最後修改時間
          format: date-time
    PackageRequest:
      required:
      - packageName
      - packageNumber
      - userAddress
      type: object
      properties:
        userAddress:
          type: string
          description: 收件人地址
        packageName:
          type: string
          description: 收件人姓名
        packageNumber:
          type: string
          description: 包裹編號
    Package:
      type: object
      properties:
        packageId:
          type: integer
          description: 包裹 ID
          format: int32
        userAddress:
          type: string
          description: 收件人地址
        packageName:
          type: string
          description: 收件人姓名
        packageNumber:
          type: string
          description: 包裹編號
        packageState:
          type: string
          description: 包裹狀態(0:未領取，1:已領取)
        createdDate:
          type: string
          description: 包裹數據建立時間
          format: date-time
        lastModifiedDate:
          type: string
          description: 包裹數據最後修改時間
          format: date-time
...